feat(db): widen account.version to bigint for the ordering guard
`account.version` is the optimistic guard for FR-10: it holds the
`provider_sequence` of the newest event applied to the account, so a snapshot
whose sequence does not exceed it arrived late and is superseded.

That makes it a store of a provider sequence, and webhook_event.provider_sequence
is bigint. As integer it truncates above 2^31-1 -- and providers that number
events from a global counter reach that scale. The failure would not be a clean
overflow error: the guard would compare against a wrong number and start
discarding live snapshots as stale.

A separate migration rather than an edit to 0001: 0001 is already in an open PR
under review, and rewriting reviewed history to avoid one four-line migration is
the wrong trade.

+++ src/webhook_receiver/adapters/orm.py
@@ -235,7 +235,15 @@ class Account(Base):
     id: Mapped[int] = _pk()
     external_ref: Mapped[str] = mapped_column(Text)
     balance_minor: Mapped[int] = mapped_column(BigInteger, server_default="0")
-    version: Mapped[int] = mapped_column(Integer, server_default="0")
+    # FR-10: the `provider_sequence` of the newest event applied to this account
+    # -- a high-water mark, not a row counter. A snapshot whose sequence does not
+    # exceed it arrived late and is superseded rather than applied.
+    #
+    # `BigInteger` because it stores a provider sequence, and
+    # `webhook_event.provider_sequence` is one. As `Integer` it would truncate
+    # above 2^31-1 and compare against a wrong number, discarding live snapshots
+    # as stale (migration 0002).
+    version: Mapped[int] = mapped_column(BigInteger, server_default="0")
     updated_at: Mapped[datetime] = mapped_column(server_default=func.now())
